scripts/requirements.txt
certifi==2020.6.20
cffi==1.14.1      
chardet==3.0.4    
cryptography==3.0
Deprecated==1.2.10
idna==2.10
pycparser==2.20
PyGithub==1.52
PyJWT==1.7.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
wrapt==1.12.1
yapf==0.30.0
